refactor: use-font-handler, use-google-fonts
