+++ .vscode/settings.json
@@ -93,6 +93,7 @@
     "unnotified",
     "unsupport",
     "Upnp",
+    "wght",
     "workarea",
     "workareas",
     "xform",
